+++ contracts/solana/programs/registry/src/errors.rs
@@ -1,0 +1,7 @@
+use anchor_lang::prelude::*;
+
+#[error_code]
+pub enum RegistryError {
+    #[msg("Schema already exists.")]
+    SchemaAlreadyExists,
+}
