+++ contracts/solana/programs/registry/src/events.rs
@@ -1,0 +1,10 @@
+use crate::state::SchemaData;
+use anchor_lang::prelude::*;
+
+#[event]
+pub struct RegisteredSchema {
+    /// The generated UID for the schema (PDA).
+    pub uid: Pubkey,
+    /// Full schema data including schema, resolver, revocable, and deployer.
+    pub schema_data: SchemaData,
+}
