+++ contracts/solana/programs/registry/src/instructions.rs
@@ -1,0 +1,3 @@
+mod register_schema;
+
+pub use register_schema::*;
