+++ contracts/solana/programs/registry/src/instructions/register_schema.rs
@@ -1,35 +1,11 @@
+use crate::errors::RegistryError;
+use crate::events::RegisteredSchema;
+use crate::state::SchemaData;
+use crate::utils::derive_schema_pda;
 use anchor_lang::prelude::*;
 
-#[event]
-pub struct RegisteredSchema {
-    /// The generated UID for the schema (PDA).
-    pub uid: Pubkey,
-    /// Full schema data including schema, resolver, revocable, and deployer.
-    pub schema_data: SchemaData,
-}
-
-#[account]
-pub struct SchemaData {
-    /// Generate PDA as reference key.
-    pub uid: Pubkey,
-    /// The actual schema data (e.g., JSON, XML, etc.).
-    pub schema: String,
-    /// Resolver address (another contract) for schema verification.
-    pub resolver: Option<Pubkey>,
-    /// Indicates whether the schema is revocable.
-    pub revocable: bool,
-    /// The deployer/authority who created the schema.
-    pub deployer: Pubkey,
-}
-
-#[error_code]
-pub enum RegistryError {
-    #[msg("Schema already exists.")]
-    SchemaAlreadyExists,
-}
-
-#[derive(Accounts)]
 /// Pass schema_name as an instruction argument.
+#[derive(Accounts)]
 #[instruction(schema_name: String)]
 pub struct RegisterSchema<'info> {
     #[account(mut)]
@@ -50,15 +26,6 @@ pub struct RegisterSchema<'info> {
     pub system_program: Program<'info, System>,
 }
 
-impl SchemaData {
-    pub const LEN: usize = 8 + 32 + 1 + 200 + 32;
-    // 8 bytes for account discriminator,
-    // 32 bytes for uid,
-    // 1 byte for revocable,
-    // 200 bytes for schema string,
-    // 32 bytes for deployer pubkey.
-}
-
 /// Registers a new schema and emits an event with the UID (PDA) and full schema data.
 ///
 /// This function creates a new schema by initializing the `SchemaData` account at the derived PDA.
@@ -75,7 +42,7 @@ impl SchemaData {
 ///
 /// # Returns
 ///
-/// * `Result<Pubkey>` - Returns the UID (PDA) of the registered schema on success.
+/// * `Result<()>` - Returns an empty Ok result on success.
 ///
 /// # Errors
 ///
@@ -94,13 +61,13 @@ impl SchemaData {
 /// Registering schemas allows users to define and store custom data structures that can be referenced
 /// in attestations. Emitting events facilitates off-chain indexing and enables clients to stay updated
 /// with new schemas without polling the blockchain.
-pub fn register_schema(
+pub fn register_schema_handler(
     ctx: Context<RegisterSchema>,
     schema_name: String,
     schema: String,
     resolver: Option<Pubkey>, // Optional resolver address for external verification.
     revocable: bool,
-) -> Result<Pubkey> {
+) -> Result<()> {
     let schema_data = &mut ctx.accounts.schema_data;
 
     // Check if the schema already exists by verifying if 'deployer' is set.
@@ -130,39 +97,7 @@ pub fn register_schema(
         }
     });
 
-    Ok(uid)
-}
+    msg!("Registered schema with UID: {:?}", uid);
 
-/// Derives the Program Derived Address (PDA) for a schema.
-///
-/// This function calculates the PDA for a schema based on the deployer's public key,
-/// the schema name, and the program ID. This PDA is used as the unique identifier
-/// (UID) for the schema and serves as the address where the schema data is stored.
-///
-/// # Arguments
-///
-/// * `deployer` - A reference to the public key of the deployer who creates the schema.
-/// * `schema_name` - The name of the schema, used as a seed in the PDA derivation.
-/// * `program_id` - A reference to the program ID of the schema registry program.
-///
-/// # Returns
-///
-/// A tuple containing:
-/// * `Pubkey` - The derived PDA for the schema.
-/// * `u8` - The bump seed associated with the PDA.
-///
-/// # Why We Are Doing This
-///
-/// Deriving a PDA ensures a unique and deterministic address for each schema based on the deployer
-/// and schema name. This prevents collisions and allows anyone to derive the address of a schema
-/// without querying the blockchain.
-pub fn derive_schema_pda(
-    deployer: &Pubkey,
-    schema_name: String,
-    program_id: &Pubkey,
-) -> (Pubkey, u8) {
-    Pubkey::find_program_address(
-        &[b"schema", deployer.as_ref(), schema_name.as_bytes()],
-        program_id,
-    )
+    Ok(())
 }
